One approach to dynamic class instantiation
The application is orchestrating things in a way that I don’t think is
optimum but to simplify things and answer what I think the original
question is about, the change in this commit may be reasonable.

What I think is at least part of the original question is how can a
plugin create an instance of a class that it can’t know about at
compile time.  For example, how can a plugin create an instance of a
class that is provided by the containing application.  In order to do
that, somehow the plugin needs to be able to discover at runtime which
class to create an instance of.  There are a number of ways this can be
done.  This commit involves introducing support for the application to
provide the name of a class in the configuration and the plugin can
retrieve that value and use it.

This commit leaves out reasonable error handling as that would only
complicate the discussion.   This code assumes that the appropriate
property is defined in application.yml.